+++ examples/nlp/duplex_text_normalization/duplex_text_normalization_train.py
@@ -5,7 +5,7 @@ from nemo.utils import logging
 from nemo.core.config import hydra_runner
 from nemo.utils.exp_manager import exp_manager
 
-@hydra_runner(config_path="conf", config_name="text_normalization_config")
+@hydra_runner(config_path="conf", config_name="duplex_tn_config")
 def main(cfg: DictConfig) -> None:
     logging.info(f'Config Params: {OmegaConf.to_yaml(cfg)}')
